I added 'update' and 'create' actions to the before action 'set_allowed_file_types'.
The reason for adding 'update' action it happens when the user
tries to upload a file that doesn't match the criteria we have to
accept such file type, then when back on the view our instance object was nil.

Same applies for the 'create' action, and the file being the wrong type.